refactor(cli): decompose doctor.rs into a doctor/ module tree
